Library Management OOP Exercise
Simple library management program. Admin and student users have different menu. 

Data of library is saved, but since it is saved in a text file, it is saved and insecure. It can be improved using openssl etc. however since it is a OOP practice, I will probably not continue on that.